Refactor lao nong logic
Summary:
主要改动有
- SystemMessage 增加 province, city, county, district, community，可以针对全国任何地区的用户。比如要显示给所有河南省的用户，只要<河南, null, null, null, null>，要现实给闸北区就<上海，上海，上海，闸北, null>
- 把LaoNong的数据逻辑，封装到ILaoNongDataSource里，为了代码清晰

    LaoNongManager manager = new LaoNongManager();
    manager.addSource(new WeatherAlertDataSource());
    manager.addSource(new SystemMessageDataSource(smDao, locDao));
    manager.addSource(new ProverbDataSource(proverbDao));
    List<LaoNong> laonongs = manager.genLaoNongs(user, areaId);

- 把所有id改成long type的。。因为IEntityDao默认所有id都必须是long type的

Test Plan:
http://localhost:8080/ploughHelper/laonongs?monitorLocationId=571400

http://localhost:8080/ploughHelper/loadWeather?lat=40.21015883788723&lon=116.649248766011&monitorLocationId=

Reviewers: Seanjiang, gaoxh04

Reviewed By: gaoxh04

Differential Revision: http://phabricator.dasinong666.com/D92

+++ src/main/java/com/dasinong/ploughHelper/ShortCutController.java
@@ -66,7 +66,7 @@ public class ShortCutController extends BaseController {
 		double lon = requestX.getDouble("lon");
 
 		try {
-			Integer mlId = AllMonitorLocation.getInstance().getNearest(lat, lon);
+			Long mlId = AllMonitorLocation.getInstance().getNearest(lat, lon);
 			result.put("respCode", 200);
 			result.put("message", "获得监控地址成功");
 			result.put("data", mlId);
