+++ src/main/java/com/dasinong/ploughHelper/dao/IMonitorLocationDao.java
@@ -6,6 +6,6 @@ import com.dasinong.ploughHelper.model.MonitorLocation;
 
 public interface IMonitorLocationDao extends IEntityDao<MonitorLocation> {
 
-	public abstract MonitorLocation findByCode(int code);
+	public abstract MonitorLocation findByCode(long code);
 
 }
