+++ src/main/java/com/dasinong/ploughHelper/facade/FieldFacade.java
@@ -123,7 +123,7 @@ public class FieldFacade implements IFieldFacade {
 
 		double lat = location.getLatitude();
 		double lon = location.getLongtitude();
-		int monitorLocationId = AllMonitorLocation.getInstance().getNearest(lat, lon);
+		long monitorLocationId = AllMonitorLocation.getInstance().getNearest(lat, lon);
 		field.setMonitorLocationId(monitorLocationId);
 		fd.save(field);
